Small refactoring and firebaseModel (#3)
* NPM Init

* Small refactoring
