+++ my-app/src/pages/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App.jsx'
+import '../index.css'
+import App from '../views/App.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
